+++ catalystwan/models/policy/list/geo_location.py
@@ -1,8 +1,7 @@
 from typing import List, Literal
 
-from catalystwan.models.policy.lists import PolicyListBase
 from catalystwan.models.policy.lists_entries import GeoLocationListEntry
-from catalystwan.models.policy.policy_list import PolicyListId, PolicyListInfo
+from catalystwan.models.policy.policy_list import PolicyListBase, PolicyListId, PolicyListInfo
 
 
 class GeoLocationList(PolicyListBase):
