+++ catalystwan/models/policy/list/port.py
@@ -1,8 +1,7 @@
 from typing import List, Literal
 
-from catalystwan.models.policy.lists import PolicyListBase
 from catalystwan.models.policy.lists_entries import PortListEntry
-from catalystwan.models.policy.policy_list import PolicyListId, PolicyListInfo
+from catalystwan.models.policy.policy_list import PolicyListBase, PolicyListId, PolicyListInfo
 
 
 class PortList(PolicyListBase):
