+++ catalystwan/models/policy/list/protocol_name.py
@@ -1,8 +1,7 @@
 from typing import List, Literal
 
-from catalystwan.models.policy.lists import PolicyListBase
 from catalystwan.models.policy.lists_entries import ProtocolNameListEntry
-from catalystwan.models.policy.policy_list import PolicyListId, PolicyListInfo
+from catalystwan.models.policy.policy_list import PolicyListBase, PolicyListId, PolicyListInfo
 
 
 class ProtocolNameList(PolicyListBase):
