+++ catalystwan/models/policy/list/vpn.py
@@ -1,8 +1,7 @@
 from typing import List, Literal, Set, Tuple
 
-from catalystwan.models.policy.lists import PolicyListBase
 from catalystwan.models.policy.lists_entries import VPNListEntry
-from catalystwan.models.policy.policy_list import PolicyListId, PolicyListInfo
+from catalystwan.models.policy.policy_list import PolicyListBase, PolicyListId, PolicyListInfo
 
 
 class VPNList(PolicyListBase):
