+++ catalystwan/models/policy/list/zone.py
@@ -1,9 +1,8 @@
 from typing import List, Literal, Set
 
 from catalystwan.models.common import InterfaceType
-from catalystwan.models.policy.lists import PolicyListBase
 from catalystwan.models.policy.lists_entries import ZoneListEntry
-from catalystwan.models.policy.policy_list import PolicyListId, PolicyListInfo
+from catalystwan.models.policy.policy_list import PolicyListBase, PolicyListId, PolicyListInfo
 
 
 class ZoneList(PolicyListBase):
