--- catalystwan/models/policy/lists.py
@@ -1,21 +1,0 @@
-# Copyright 2023 Cisco Systems, Inc. and its affiliates
-
-from typing import Any, List, Optional
-
-from pydantic import BaseModel, Field
-
-
-class PolicyListBase(BaseModel):
-    name: str = Field(
-        pattern="^[a-zA-Z0-9_-]{1,32}$",
-        description="Can include only alpha-numeric characters, hyphen '-' or underscore '_'; maximum 32 characters",
-    )
-    description: Optional[str] = "Desc Not Required"
-    entries: List[Any]
-
-    def _add_entry(self, entry: Any, single: bool = False) -> None:
-        if self.entries and single:
-            self.entries[0] = entry
-            del self.entries[1:]
-        else:
-            self.entries.append(entry)
